MAINT: Move _inspect and _pep440 from compat to _utils
Note that unfortunately, compat does expose _inspect as well,
so the import remains (just the definition place moves).
