+++ numpy/_utils/_inspect.py
