+++ numpy/compat/__init__.py
@@ -8,9 +8,10 @@ extensions, which may be included for the following reasons:
   * we may only need a small subset of the copied library/module
 
 """
-from . import _inspect
+
+from .._utils import _inspect
+from .._utils._inspect import getargspec, formatargspec
 from . import py3k
-from ._inspect import getargspec, formatargspec
 from .py3k import *
 
 __all__ = []
